requirements.txt
groq>=0.4.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.22
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
numpy>=1.24.3
